refactor: dynamic Namespace 적용 및 sockets 디렉토리 이동
- 소켓 이벤트 설정 코드 sockets/로 분리
- dynamic Namespace로 uuid 형식 roomId 받아 소켓 연결
